Simplify run-stunnel script
Most of the run-stunnel script is also present in the start-stunnel
script.  This change reuses this code.

It also ensures that the aux (i.e. stunnel) job that is created by both scripts
is terminated.  It first  attempts to send a SIGINT (as it did previously),
however if after 3 seconds the process has not terminated it instead
sends a SIGKILL.

This ensures that there are not dangling stunnel processes left over after the
script is run and that repeated executions of the script will have predicatable
results.